pyproject.toml
[project]
name = "mie-bot"
version = "0.1.0"
description = "abcxyz"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "async-cache>=2.0.0",
    "beanie>=2.0.1",
    "cachetools>=7.0.5",
    "discord.py>=2.4.0",
    "google-genai>=1.66.0",
    "motor>=3.7.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytz>=2026.1.post1",
    "quickchart-io>=2.0.0",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F405"]
"main.py" = ["F403"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
